requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

discord.py==2.3.2
python-dotenv==1.0.1
requests==2.31.0
